refactor: make MultiThreaded the default mode
In multithreaded mode, rocksdb uses its own locking to ensure thread-safe access to the DB:
https://docs.rs/rocksdb/latest/rocksdb/trait.ThreadMode.html

This was the intended use for this API, as evidence by storing the DB reference in `Arc`, but
the DB was opened everywhere in SingleThreaded mode, as by default.

This was a bug that the present PR correcs.