fix: Make the Psk structure allocating
The Psk structure was a non-allocating structure (creating reference from a reference), due to the non-allocating style of the parser in the X509 parser.

This is unwieldy for several reasons, not the least because using a ClientCertVerifier of a ServerCertVerifier in the rustls framework requires either the ability to pass a 'static reference to the underlying value.
This forgoes the theatrics, and makes Psk allocating by explicilty allocatingthe SPKI bytes it represents.